requirements.txt
Flask==2.2.3
Flask-SocketIO==5.3.4
openai==0.28.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
requests==2.21.0